refactor: BackendConfig trait, native-dtype TensorStore, fix eps scalar type
Items A-F from backend-generics-abstraction analysis:
- Item A: norm.rs add_scalar(eps as f32) -> add_scalar(eps) (f64 pass-through)
- Item B: new src/backend_config.rs with BackendConfig trait and impls for all
  6 concrete backends (NdArray, Wgpu, Cuda f32/bf16, LibTorch f32/bf16)
- Item C: infer.rs expanded backend selection to all 6 backends + compile_error
  guards for all invalid combos
- Item E: TensorStore refactored from HashMap<String, Vec<f32>> to
  HashMap<String, TensorEntry> with raw bytes + dtype + shape; native-dtype
  decoding via TensorEntry::to_tensor_data<D>(); byte-length validation
- Item F: --gpu-id u32 CLI arg added to bench_realmodel and infer; shown in output
- All binaries updated to use BackendConfig::device_from_id/backend_label

Results after refactor:
- E2E correctness: PASS (max_abs_diff=0.0)
- LibTorch bf16: 1,899ms (was 1,939ms, slight improvement from native-dtype load)
- Wgpu: 7,033ms (was 7,396ms, 5% improvement from native-dtype load)
- benchmarks.md updated with new measurements

+++ src/model/norm.rs
@@ -25,7 +25,7 @@ impl<B: Backend> RmsNorm<B> {
     pub fn forward(&self, x: Tensor<B, 3>) -> Tensor<B, 3> {
         let rms = (x.clone() * x.clone())
             .mean_dim(2) // [B, S, 1] (keepdim)
-            .add_scalar(self.eps as f32)
+            .add_scalar(self.eps)
             .sqrt(); // [B, S, 1]
         // weight: [D] broadcasts to [1, 1, D] via burn's automatic broadcasting
         let w: Tensor<B, 3> = self
@@ -58,7 +58,7 @@ impl<B: Backend> HeadRmsNorm<B> {
     pub fn forward(&self, x: Tensor<B, 4>) -> Tensor<B, 4> {
         let rms = (x.clone() * x.clone())
             .mean_dim(3) // [B, S, H, 1]
-            .add_scalar(self.eps as f32)
+            .add_scalar(self.eps)
             .sqrt(); // [B, S, H, 1]
         // weight: [H, D_h] → [1, 1, H, D_h]
         let w: Tensor<B, 4> = self
@@ -154,7 +154,7 @@ impl<B: Backend> LowRankAdaLn<B> {
         // RMSNorm x: [B, S, D]
         let rms = (x.clone() * x.clone())
             .mean_dim(2)
-            .add_scalar(self.eps as f32)
+            .add_scalar(self.eps)
             .sqrt(); // [B, S, 1]
         let x_norm = x / rms;
